refactor: moved account to view folder
